refactor: use shared formatters for file sizes and timestamps in index.tsx
- Replace local inline formatFileSize and formatTimestamp with centralized utilities
- Import formatFileSize and formatTimestamp from utils/formatters
- Remove obsolete Japanese inline comments and helper functions
- Improve consistency and reuse across codebase
