+++ projects/file-server/src/utils/formatters.ts
@@ -1,0 +1,19 @@
+export function formatFileSize(bytes: number): string {
+  if (bytes === 0) return "0 B"
+  const k = 1024
+  const sizes = ["Byte", "KB", "MB", "GB", "TB"]
+  const i = Math.floor(Math.log(bytes) / Math.log(k))
+  return `${parseFloat((bytes / k ** i).toFixed(2))} ${sizes[i]}`
+}
+
+const pad2 = (n: number) => n.toString().padStart(2, "0")
+
+export function formatTimestamp(date: Date): string {
+  const y = date.getFullYear()
+  const mo = pad2(date.getMonth() + 1)
+  const d = pad2(date.getDate())
+  const h = pad2(date.getHours())
+  const mi = pad2(date.getMinutes())
+  const s = pad2(date.getSeconds())
+  return `${y}-${mo}-${d} ${h}:${mi}:${s}`
+}
